knowledge-admin/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.3.3
openpyxl==3.1.5
openai==1.54.0
httpx==0.27.0
# 認證相關套件
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
